requirements.txt
# Core Deep Learning Libraries (GPU - CUDA 12.6)
torch==2.6.0+cu126
torchvision==0.21.0+cu126
torchaudio==2.6.0+cu126
--extra-index-url https://download.pytorch.org/whl/cu126

# YOLOv8
ultralytics==8.3.25

# Face Recognition (DeepFace)
deepface==0.0.93

# Vector Database (FAISS)
faiss-cpu==1.8.0

# Image & Video Processing
opencv-python==4.10.0.84
Pillow==10.4.0

# Data Handling
numpy==1.26.4
pandas==2.2.3
pickle-mixin==1.0.2


tqdm==4.66.5
matplotlib==3.9.2


pyyaml==6.0.2
